Values of setfeetconstraint are too large. 4cm of security margin is a good value
This problem did not appear because of an error in JRL-walkgen.
The value is supposed to be used in the method
RelativeFeetInequalities::CallMethod, but in previous versions, the convex hull was not updated. Hence, the given value was not taken into account.
The pg worked because the default value (4cm) was correct.
This problem has only be corrected in the branch topic/rc-v3.1.4 (commit 54fd8547ee47...)